feat(pkg/scanner): add parseLinks utility
This new function will be used to parse the links inside an HTML body
from a request response. The idea is that it will tokenize the contents
of an HTTP response using HTML rules, pulling out the links contained on
the `href` property of the `a` tag.

Also added the utility functions `isBrowsableURL` (to check if an URL is
browsable), `makeAbsoluteURL` (to convert relative URLs into absolute
URLs). Added tests for both functions.