feat(pkg/scanner): improve parseLinks reliability
Add parametrization to the function so the behaviour can be configured.
Add graceful handling of nil response bodies, returns empty slice if that
is the case.
Add maximum read limits to the function to prevent high memory usage.

BREAKING CHANGE: function interface has changed, needing the
DefaultLinkParserOptions param either to be passed or set to nil